refactor(pull_reqeust): :recycle: moved close pull request label to another file
