+++ index.js
@@ -4,7 +4,7 @@ const {
 	addMergedLabel,
 	WIPLabelAutomation: pullRequestWIPLabelAutomation,
 	changesRequestLabel,
-	removeWIPLabel,
+	addCloseLabel,
 } = require("./automation/addLabelsOnPullRequest.automation");
 const addLabelToIssueOnClose = require("./automation/addLabelToIssueOnClose.automation");
 const approveCommand = require("./commands/approve.command");
@@ -36,40 +36,7 @@ module.exports = app => {
 	app.on("pull_request_review.submitted", changesRequestLabel);
 
 	app.on("pull_request.closed", addMergedLabel);
-	app.on("pull_request.closed", async context => {
-		const params = context.pullRequest();
-
-		try {
-			const pullRequestIsMerged = await context.octokit.pulls.checkIfMerged(
-				params
-			);
-		} catch (err) {
-			const issueLabels = await listIssueLabels(context);
-			const foundReadyForReviewLabel = issueLabels.data.filter(
-				label => label.name === ":mag: Ready for Review"
-			);
-			const foundApprovedLabel = issueLabels.data.filter(
-				label => label.name === ":white_check_mark: Approved"
-			);
-			const foundChangesRequestedLabel = issueLabels.data.filter(
-				label => label.name === ":warning: Changes requested"
-			);
-
-			if (foundReadyForReviewLabel.length > 0) {
-				removeLabel([":mag: Ready for Review"], context);
-			}
-
-			if (foundApprovedLabel.length > 0) {
-				removeLabel([":white_check_mark: Approved"], context);
-			}
-
-			if (foundChangesRequestedLabel.length > 0) {
-				removeLabel([":warning: Changes requested"], context);
-			}
-
-			addLabel([":x: closed"], context, "B60205");
-		}
-	});
+	app.on("pull_request.closed", addCloseLabel);
 
 	app.on(
 		["pull_request.edited", "pull_request.labeled"],
